+++ src/ManagementPartner/ManagementPartner/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Removed Microsoft.Azure.Management.ManagementPartner 1.1.1-preview dependency
+* Added Microsoft.Azure.PowerShell.ManagementPartner.Management.Sdk
 
 ## Version 0.7.3
 * Refreshed module to ensure catalog file signed by Microsoft.